Refactor UserConfigForm to own component
